requirements.txt
vk_api
mysql-connector-python